Remove extensions/v1beta1 Job
