+++ pkg/client/clientset_generated/release_1_3/typed/extensions/v1beta1/extensions_client.go
@@ -29,7 +29,6 @@ type ExtensionsInterface interface {
 	DeploymentsGetter
 	HorizontalPodAutoscalersGetter
 	IngressesGetter
-	JobsGetter
 	PodSecurityPoliciesGetter
 	ReplicaSetsGetter
 	ScalesGetter
@@ -55,10 +54,6 @@ func (c *ExtensionsClient) HorizontalPodAutoscalers(namespace string) Horizontal
 
 func (c *ExtensionsClient) Ingresses(namespace string) IngressInterface {
 	return newIngresses(c, namespace)
-}
-
-func (c *ExtensionsClient) Jobs(namespace string) JobInterface {
-	return newJobs(c, namespace)
 }
 
 func (c *ExtensionsClient) PodSecurityPolicies() PodSecurityPolicyInterface {
