+++ pkg/client/clientset_generated/release_1_5/typed/extensions/v1beta1/generated_expansion.go
@@ -20,8 +20,6 @@ type DaemonSetExpansion interface{}
 
 type IngressExpansion interface{}
 
-type JobExpansion interface{}
-
 type PodSecurityPolicyExpansion interface{}
 
 type ReplicaSetExpansion interface{}
