+++ pkg/client/unversioned/testclient/fake_jobs.go
@@ -23,9 +23,10 @@ import (
 )
 
 // FakeJobs implements JobInterface. Meant to be embedded into a struct to get a default
-// implementation. This makes faking out just the method you want to test easier.
+// implementation. This makes faking out just the methods you want to test easier.
+// This is a test implementation of JobsV1
 type FakeJobs struct {
-	Fake      *FakeExperimental
+	Fake      *FakeBatch
 	Namespace string
 }
 
@@ -75,69 +76,6 @@ func (c *FakeJobs) Watch(opts api.ListOptions) (watch.Interface, error) {
 }
 
 func (c *FakeJobs) UpdateStatus(job *batch.Job) (result *batch.Job, err error) {
-	obj, err := c.Fake.Invokes(NewUpdateSubresourceAction("jobs", "status", c.Namespace, job), job)
-	if obj == nil {
-		return nil, err
-	}
-
-	return obj.(*batch.Job), err
-}
-
-// FakeJobs implements JobInterface. Meant to be embedded into a struct to get a default
-// implementation. This makes faking out just the methods you want to test easier.
-// This is a test implementation of JobsV1
-// TODO(piosz): get back to one client implementation once HPA will be graduated to GA completely
-type FakeJobsV1 struct {
-	Fake      *FakeBatch
-	Namespace string
-}
-
-func (c *FakeJobsV1) Get(name string) (*batch.Job, error) {
-	obj, err := c.Fake.Invokes(NewGetAction("jobs", c.Namespace, name), &batch.Job{})
-	if obj == nil {
-		return nil, err
-	}
-
-	return obj.(*batch.Job), err
-}
-
-func (c *FakeJobsV1) List(opts api.ListOptions) (*batch.JobList, error) {
-	obj, err := c.Fake.Invokes(NewListAction("jobs", c.Namespace, opts), &batch.JobList{})
-	if obj == nil {
-		return nil, err
-	}
-
-	return obj.(*batch.JobList), err
-}
-
-func (c *FakeJobsV1) Create(job *batch.Job) (*batch.Job, error) {
-	obj, err := c.Fake.Invokes(NewCreateAction("jobs", c.Namespace, job), job)
-	if obj == nil {
-		return nil, err
-	}
-
-	return obj.(*batch.Job), err
-}
-
-func (c *FakeJobsV1) Update(job *batch.Job) (*batch.Job, error) {
-	obj, err := c.Fake.Invokes(NewUpdateAction("jobs", c.Namespace, job), job)
-	if obj == nil {
-		return nil, err
-	}
-
-	return obj.(*batch.Job), err
-}
-
-func (c *FakeJobsV1) Delete(name string, options *api.DeleteOptions) error {
-	_, err := c.Fake.Invokes(NewDeleteAction("jobs", c.Namespace, name), &batch.Job{})
-	return err
-}
-
-func (c *FakeJobsV1) Watch(opts api.ListOptions) (watch.Interface, error) {
-	return c.Fake.InvokesWatch(NewWatchAction("jobs", c.Namespace, opts))
-}
-
-func (c *FakeJobsV1) UpdateStatus(job *batch.Job) (result *batch.Job, err error) {
 	obj, err := c.Fake.Invokes(NewUpdateSubresourceAction("jobs", "status", c.Namespace, job), job)
 	if obj == nil {
 		return nil, err
